requirements.txt
fastapi[all]
sqlalchemy
asyncpg
psycopg2-binary
python-dotenv
authlib
passlib
pyjwt
starlette
uvicorn
